--- packages/umi-utils/src/endWithSlash.js
@@ -1,3 +1,0 @@
-export default function endWithSlash(path) {
-  return path.slice(-1) !== '/' ? `${path}/` : path;
-}
