--- packages/umi-utils/src/endWithSlash.test.js
@@ -1,8 +1,0 @@
-import endWithSlash from './endWithSlash';
-
-describe('endWithSlash', () => {
-  test('normal', () => {
-    expect(endWithSlash('./a')).toEqual('./a/');
-    expect(endWithSlash('./a/')).toEqual('./a/');
-  });
-});
