+++ packages/umi-utils/src/endWithSlash.test.ts
@@ -1,0 +1,6 @@
+import endWithSlash from './endWithSlash';
+
+test('normal', () => {
+  expect(endWithSlash('./a')).toEqual('./a/');
+  expect(endWithSlash('./a/')).toEqual('./a/');
+});
