+++ packages/umi-utils/src/findJS.ts
@@ -3,7 +3,12 @@ import { existsSync } from 'fs';
 
 const JS_EXTNAMES = ['.js', '.jsx', '.ts', '.tsx'];
 
-export default function(baseDir, fileNameWithoutExtname) {
+/**
+ * Find the real JS file. Automatic completion of suffixes
+ * @param baseDir base path
+ * @param fileNameWithoutExtname file name
+ */
+export default function(baseDir: string, fileNameWithoutExtname: string) {
   let i = 0;
   while (i < JS_EXTNAMES.length) {
     const extname = JS_EXTNAMES[i];
