+++ packages/umi-utils/src/index.ts
@@ -1,0 +1,11 @@
+export { default as findJS } from './findJS';
+export { default as compatDirname } from './compatDirname';
+export { default as winPath } from './winPath';
+export { default as optsToArray } from './optsToArray';
+export { default as chunkName } from './chunkName';
+export { default as endWithSlash } from './endWithSlash';
+export { default as isUrl } from './isUrl';
+export { default as deprecate } from './deprecate';
+export { default as winEOL } from './winEOL';
+export { default as prettierFile } from './prettierFile';
+export { default as getFastGithub } from './getFastGithub';
