--- packages/umi-utils/src/isUrl.js
@@ -1,5 +1,0 @@
-import isUrl from 'is-url';
-
-export default function(path) {
-  return isUrl(path);
-}
