+++ packages/umi-utils/src/isUrl.ts
@@ -1,0 +1,9 @@
+import isUrl from 'is-url';
+
+/**
+ * Check whether a string is a URL.
+ * @param path
+ */
+export default function(path: string) {
+  return isUrl(path);
+}
