--- packages/umi-utils/src/optsToArray.js
@@ -1,8 +1,0 @@
-export default function optsToArray(item) {
-  if (item === null || item === undefined) return [];
-  if (Array.isArray(item)) {
-    return item;
-  } else {
-    return [item];
-  }
-}
