+++ packages/umi-utils/src/optsToArray.test.ts
