+++ packages/umi-utils/src/winPath.ts
@@ -1,0 +1,9 @@
+import slash from 'slash2';
+
+/**
+ * Convert Windows backslash paths to slash paths
+ * @param path
+ */
+export default function(path: string) {
+  return slash(path);
+}
